requirements.txt
pillow==10.0.0
tflite-runtime==2.13.0
discord.py==2.3.2
